PID update to ATrack
The PID tracking correction function has now been added to ATrack.  To implement this a new tab (PID) has been added to the GUI and a new pascal unit (PID.pas) has been added.

A PID checkbox, located on the PID tab, is used to select which algorithm is used for correction.